+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/pipelines/processors/ResumeStateStoringProcessor.java
@@ -24,7 +24,7 @@ import com.walmartlabs.concord.project.InternalConstants;
 import com.walmartlabs.concord.server.process.Payload;
 import com.walmartlabs.concord.server.process.ProcessException;
 import com.walmartlabs.concord.server.process.ProcessKey;
-import com.walmartlabs.concord.server.process.logs.LogManager;
+import com.walmartlabs.concord.server.process.logs.ProcessLogManager;
 
 import javax.inject.Inject;
 import javax.inject.Named;
@@ -36,10 +36,10 @@ import java.nio.file.Path;
 public class ResumeStateStoringProcessor implements PayloadProcessor {
 
 
-    private final LogManager logManager;
+    private final ProcessLogManager logManager;
 
     @Inject
-    public ResumeStateStoringProcessor(LogManager logManager) {
+    public ResumeStateStoringProcessor(ProcessLogManager logManager) {
         this.logManager = logManager;
     }