+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/pipelines/processors/policy/WorkspacePolicyApplier.java
@@ -27,7 +27,7 @@ import com.walmartlabs.concord.policyengine.WorkspaceRule;
 import com.walmartlabs.concord.server.process.Payload;
 import com.walmartlabs.concord.server.process.ProcessException;
 import com.walmartlabs.concord.server.process.ProcessKey;
-import com.walmartlabs.concord.server.process.logs.LogManager;
+import com.walmartlabs.concord.server.process.logs.ProcessLogManager;
 import com.walmartlabs.concord.server.sdk.metrics.InjectCounter;
 
 import javax.inject.Inject;
@@ -39,7 +39,7 @@ import static com.walmartlabs.concord.server.process.pipelines.processors.policy
 @Named
 public class WorkspacePolicyApplier implements PolicyApplier {
 
-    private final LogManager logManager;
+    private final ProcessLogManager logManager;
 
     @InjectCounter
     private final Counter policyWarn;
@@ -48,7 +48,7 @@ public class WorkspacePolicyApplier implements PolicyApplier {
     private final Counter policyDeny;
 
     @Inject
-    public WorkspacePolicyApplier(LogManager logManager, Counter policyWarn, Counter policyDeny) {
+    public WorkspacePolicyApplier(ProcessLogManager logManager, Counter policyWarn, Counter policyDeny) {
         this.logManager = logManager;
         this.policyWarn = policyWarn;
         this.policyDeny = policyDeny;